\name{mosaicManip-package}
\alias{mosaicManip-package}
\alias{mosaicManip}
\docType{package}
\title{
	Example manipulate applets
}
\description{
	Demonstration applets for Statistics, Calculus, and Epidemiology
	using the functionality in the \code{manipulate} package.
}
\details{
\tabular{ll}{
Package: \tab mosaicManip\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2011-06-23\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel Kaplan

Maintainer: Daniel Kaplan
}
\seealso{
code{\link[manipulate:manipulate-package]{manipulate}} 
}
\keyword{ package }
