\name{undocumented}
\docType{methods}

\alias{mEulerIntegrate}
\alias{mGlassPattern}
\alias{mGradSearch}
\alias{mLM}
\alias{mLinAlgebra}
\alias{mLineFit}
\alias{mPhaseLine}
\alias{mhistogram}
\alias{mpoly2}

\title{Undocumented Objects}

\description{Documentation is still pending for these objects}

\author{Randall Pruim, Daniel Kaplan}

\examples{
}
