\name{mosaicManip-package}
\alias{mosaicManip-package}
\alias{mosaicManip}
\docType{package}
\title{
	Example manipulate applets
}
\description{
	Demonstration applets for Statistics, Calculus, and Epidemiology
	using the functionality in the \code{manipulate} package.
}
\details{
\tabular{ll}{
Package: \tab mosaicManip\cr
Type: \tab Package\cr
Version: \tab 0.2-4\cr
Date: \tab 2011-12-28\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Daniel Kaplan, Andrew Rich, Randall Pruim

Maintainer: Daniel Kaplan, Randall Pruim
}
%\seealso{
%code{\link[manipulate:manipulate-package]{manipulate}} 
%}
\keyword{ package }
